Rename astroid.constants to astroid.const
See https://github.com/PyCQA/astroid/pull/1045\#discussion_r654572722
