+++ astroid/const.py
