+++ astroid/protocols.py
@@ -32,7 +32,7 @@ import operator as operator_mod
 from astroid import arguments, bases
 from astroid import context as contextmod
 from astroid import decorators, helpers, node_classes, nodes, util
-from astroid.constants import Store
+from astroid.const import Store
 from astroid.exceptions import (
     AstroidIndexError,
     AstroidTypeError,
